bgp: fix BFD CRD status fields and vendored GoBGP state mapping
Three bugs fixed in the BFD status reporting pipeline:

Bug A - Missing BFD fields in toConfig(): The vendored toConfig() in
server.go only copied SessionState, FailureTransitions, LocalDiscriminator,
and BfdAsync into the OC struct. Added RemoteSessionState,
RemoteDiscriminator, LocalDiagnosticCode, and RemoteDiagnosticCode. Also
added apiBfdDiagnosticCodeToOC() helper using oc.IntToBfdDiagnosticCodeMap.

Bug B - CRD type mismatch: LocalDiscriminator/RemoteDiscriminator were
*uint32 and FailureTransitions was *uint64, but CRD schema generated
format:int32 which rejected values > 2147483647 (worker node discriminator
was 3896274385). Changed all three to *int64 in bgp_node_types.go,
crd_status.go, and CRD YAML.

Bug C - Remote session state enum mismatch: p.state stores
api.BfdSessionState values (direct cast OK for local), but
p.remoteSessionState stores bfd.StateType values from wire packets.
bfd.StateUp(3) maps to api.BFD_SESSION_STATE_UP(1), not ADMIN_DOWN(3).
Added bfdStateToAPIState() mapping function for remote state. Also added
deriveLocalDiagnostic() using api.BfdSessionState switch. Added atomic
fields (remoteSessionState, remoteDiagnosticCode, failureTransitions) to
bfdPeer struct and populates them in rxPacket/setStateDown.

Also: regenerated deepcopy, extended state_test.go assertions, updated
containerlab bgp.yaml BFD field names to match new API.

+++ pkg/bgp/manager/reconciler/crd_status.go
@@ -430,9 +430,9 @@ func (r *StatusReconciler) getInstanceStatus(ctx context.Context, instance *inst
 				RemoteSessionState:  ptr.To[string](bfd.RemoteSessionState),
 				LocalDiagnosticCode: ptr.To[string](bfd.LocalDiagnosticCode),
 				RemoteDiagnosticCode: ptr.To[string](bfd.RemoteDiagnosticCode),
-				LocalDiscriminator:  ptr.To[uint32](bfd.LocalDiscriminator),
-				RemoteDiscriminator: ptr.To[uint32](bfd.RemoteDiscriminator),
-				FailureTransitions:  ptr.To[uint64](bfd.FailureTransitions),
+				LocalDiscriminator:  ptr.To[int64](int64(bfd.LocalDiscriminator)),
+				RemoteDiscriminator: ptr.To[int64](int64(bfd.RemoteDiscriminator)),
+				FailureTransitions:  ptr.To[int64](int64(bfd.FailureTransitions)),
 			}
 		}
